backend/requirements-min.txt
fastapi[standard]
sqlalchemy
psycopg2-binary
python-jose[cryptography]
argon2-cffi
